Add 3 new functions: check_packet, check_uri function and check_headers
check_packet parses through the sip packet and searches for mandatory
headers if the packet is a request for now.

it does so by calling the check_headers function on them which detects
if the headers are RFC3261 compliant and throws exceptions if they are
not.

the check_uri function parses the uri provided for the request-line,
to, from and contact headers and validates it.