Fix bugs in print function
Both overloads of the print function were introducing a bug where
the printed data would not be formatted properly.

Specifically the METHOD_NAME for the SIP package was being followed
by ":".
We also did not account for the fact that the normal SIP header
can also have multiple fields of the same name that need to be
printed using the method we used to print the multiple SDP fields.

To fix these bugs I added checks to differentiate between the
method name and other header field names and also added checks
for multiple same name regular header fields.