Change SIP packet output method to improve overall performance
Most of the SIP processing time was spent on writing files. The initial
setup was that each captured SIP packet got it's own output file. This
was unnecessary, and for large input sizes lead to most of the time
being spent opening and closing files.

The new solution removes this, takes a single file approach and cuts
the processing time in half.